+++ app/src/main/java/com/planzy/app/MainActivity.kt
@@ -132,6 +132,7 @@ class MainActivity : ComponentActivity() {
                     Navigation(
                         deepLinkViewModel = deepLinkViewModel,
                         navController = navController,
+                        searchViewModel = searchViewModel,
                         modifier = Modifier
                             .fillMaxSize()
                             .padding(innerPadding)
