+++ app/src/main/java/com/planzy/app/ui/screens/SearchViewModel.kt
@@ -32,6 +32,12 @@ data class PlaceWithStats(
     val userReviewsCount: Int
 )
 
+data class SearchResults(
+    val placesWithStats: List<PlaceWithStats>,
+    val vacations: List<Vacation>,
+    val users: List<User>
+)
+
 class SearchViewModel(
     private val searchPlacesUseCase: SearchPlacesUseCase,
     private val getUserCommentsStatsUseCase: GetUserCommentsStatsUseCase,
@@ -51,7 +57,7 @@ class SearchViewModel(
     }
 
     private val prefs = context.getSharedPreferences("planzy_prefs", Context.MODE_PRIVATE)
-    private val searchCache = mutableMapOf<String, List<PlaceWithStats>>()
+    private val searchCache = mutableMapOf<String, SearchResults>()
 
     var searchQuery by mutableStateOf("")
         private set
@@ -77,6 +83,8 @@ class SearchViewModel(
         private set
     var users by mutableStateOf<List<User>>(emptyList())
         private set
+
+    private var searchJob: kotlinx.coroutines.Job? = null
 
     init {
         viewModelScope.launch {
@@ -159,29 +167,38 @@ class SearchViewModel(
             vacations = emptyList()
             users = emptyList()
             errorMessage = null
+            searchJob?.cancel()
             return
         }
 
         if (searchCache.containsKey(cleanQuery)) {
-            placesWithStats = searchCache[cleanQuery]!!
-            places = placesWithStats.map { it.place }
+            val cached = searchCache[cleanQuery]!!
+            placesWithStats = cached.placesWithStats
+            places = cached.placesWithStats.map { it.place }
+            vacations = cached.vacations
+            users = cached.users
             errorMessage = null
+            searchJob?.cancel()
             return
         }
 
-        viewModelScope.launch {
+        searchJob?.cancel()
+        searchJob = viewModelScope.launch {
+            kotlinx.coroutines.delay(500)
+
             isLoading = true
             errorMessage = null
 
             Log.d(TAG, "Starting search for: $cleanQuery")
 
             val usersResult = searchUsersUseCase(cleanQuery)
-
             var hasUsers = false
+            var usersList = emptyList<User>()
             usersResult.onSuccess { list ->
                 Log.d(TAG, "Found ${list.size} users")
                 hasUsers = list.isNotEmpty()
                 users = list
+                usersList = list
             }.onFailure { exception ->
                 Log.e(TAG, "Error searching users: ${exception.message}", exception)
             }
@@ -190,6 +207,7 @@ class SearchViewModel(
             val vacationsResult = searchVacationsUseCase(cleanQuery)
 
             var hasVacations = false
+            var vacationsList = emptyList<Vacation>()
             vacationsResult.onSuccess { list ->
                 Log.d(TAG, "Found ${list.size} vacations")
                 hasVacations = list.isNotEmpty()
@@ -202,7 +220,7 @@ class SearchViewModel(
                 }
 
                 vacations = vacationsWithComments
-                Log.d(TAG, "Updated vacations with comments: ${vacationsWithComments.map { "${it.title} (${it.commentsCount} comments)" }}")
+                vacationsList = vacationsWithComments
             }.onFailure { exception ->
                 Log.e(TAG, "Error searching vacations: ${exception.message}", exception)
             }
@@ -215,6 +233,7 @@ class SearchViewModel(
             val placesResult = searchPlacesUseCase(cleanQuery, latLong = latLong, radius = radius)
 
             var hasPlaces = false
+            var placeStatsList = emptyList<PlaceWithStats>()
             placesResult.onSuccess { list ->
                 places = list
                 hasPlaces = list.isNotEmpty()
@@ -227,11 +246,19 @@ class SearchViewModel(
                 }
 
                 placesWithStats = stats
-                searchCache[cleanQuery] = stats
+                placeStatsList = stats
             }.onFailure { exception ->
                 if (!hasVacations) {
                     errorMessage = resourceProvider.getString(mapExceptionToErrorResource(exception))
                 }
+            }
+
+            if (hasPlaces || hasVacations || hasUsers) {
+                searchCache[cleanQuery] = SearchResults(
+                    placesWithStats = placeStatsList,
+                    vacations = vacationsList,
+                    users = usersList
+                )
             }
 
             if (!hasPlaces && !hasVacations && !hasUsers) {
